fix Abstract naming. Refactor
