+++ src/presenter/trip.js
