+++ src/view/event/list.js
@@ -1,6 +1,6 @@
-import Abstract from '../abstract.js';
+import AbstractView from '../abstract.js';
 
-export default class EventListView extends Abstract {
+export default class EventListView extends AbstractView {
   _getTemplate() {
     return (
       `<ul class="trip-events__list">
