+++ src/view/menu/controls.js
@@ -1,6 +1,6 @@
-import Abstract from '../abstract.js';
+import AbstractView from '../abstract.js';
 
-export default class MenuControlsView extends Abstract {
+export default class MenuControlsView extends AbstractView {
   _getTemplate() {
     return (
       `<div class="trip-main__trip-controls  trip-controls">
