+++ src/view/menu/filters.js
@@ -1,6 +1,6 @@
-import Abstract from '../abstract.js';
+import AbstractView from '../abstract.js';
 
-export default class MenuFiltersView extends Abstract {
+export default class MenuFiltersView extends AbstractView {
   _getTemplate() {
     return (
       `<h2 class="visually-hidden">Filter events</h2>
